Extend "current time in ms" to 64 bits
If we only use 32 bits to express "current time in ms" then this value
will surely loop back around every 2^32 ms, which is about 50 days.

The idea of a game of Doom running continually for 50 days sounds rare,
but if the imported `getTicksMs` has an intrinsic offset in the value it
provides (e.g. if `getTickMs` just returns "number of ms since last
reboot"), then a 32-bit representation of this value may loop around
arbitrarily soon after Doom starts up. E.g. if Doom runs for 6 hours then
there's a 1/200 chance that `getTickMs` will loop around.

We don't want to force the implementation of `getTicksMs` to not have
any implicit offset (because requiring it to return '0' or something
close to 0 the first time it's called sounds messy to manage from the
implementation side).  Instead we just want to require that `getTicksMs`
always returns a larger and larger values as time passes.

So, to make it easier for our users to not worry about the value
returned by `getTicksMs` looping around to 0 during the time Doom is
running, we'll just extend this value to 64 bits.

Given an arbitrary offset in the values returned by `getTickMs` and Doom
running for 6 hours, with 64 bits of precision we'll now only have a
1/800000000000 chance that `getTickMs` will loop around.

+++ doom_wasm.h
@@ -37,7 +37,7 @@ IMPORT_MODULE("ui") void drawFrame(uint32_t *screenBuffer);
 IMPORT_MODULE("console") void info(const char *message, size_t length);
 IMPORT_MODULE("console") void error(const char *message, size_t length);
 
-IMPORT_MODULE("runtimeControl") uint32_t getTicksMs();
+IMPORT_MODULE("runtimeControl") uint64_t getTicksMs();
 
 // Return the size of the associated save game data, 0 in the case that no save
 // data exists for this slot
